Fix bug in sumw2 handling
In coffea.hist.Hist there are basically only two options for sumw2.  For a histogram either you're keeping track of sumw2 or not.  With HistEFT this gets slightly more complicated in that there is a large cost to storing the w2 coefficients, so sometimes you want to skip that.  However, you should still track sumw2 for the bins that don't have EFT coefficients.  This commit changes the sumw2 handling so that the it's handled correctly in all cases, meaning for EFT samples with w**2 coefficients, with EFT samples without w**2 coefficients, for weighted non-EFT samples and for unweighted non-EFT samples.  This fixes the bug reported by Brent related to the scale function.