Initial Commit for Reducing Memory
This is a minimal commit for testing the approach for reducing the memory usage.  I've replaced the EFTHelper class which allocated memory for auxilary arrays and which had a separate copy stored in each HistEFT with functions that don't rely on any persistently allocated memory.  Instead, these functions use Numba and explicit loops to complete the necessary calculations elementwise, only allocating memory for the final result.