Enable ability to handle samples with different WC lists
This commit provides the ability to fill different EFT samples in the same HistEFT even if the WC lists are different.  The topeft.py processor is provided with a list of the desired WCs in its constructor, and the HistEFTs are constructed according to that list.  During the run of the processor, the actual list of WCs for the sample being processed is compared to the desired list.  The EFT weight fit coefficients are reordered to match the desired list.  The fit coefficients for any WCs on the desired list that are not present for this sample are set to zero.  If there are any WCs in the current sample that are not in the list of desired coefficients, they are dropped.

By default, the run.py and work_queue_run.py scripts set the list of desired WCs in the processor to the union of all sets of WCs for the individual samples processed.  Optionally, you can override this list by specifying your own list with the `--wc-list` command line argument.

There are two small changes that are not directly related to managing WC lists included here just to have them somewhere:
* The `--do-systs` command line argument was propagated to work_queue_run.py
* An appropriate `.gitignore` files was added to the repository to declutter the `git status` output.