Fix bugs in handling off sumw2
Thanks to @btovar, I was able to isolate two bugs in HistEFT:
 1. In some cases, adding a HistEFT with no EFT weights to one with EFT weight would result in the `_sumw` values being erroneously copied into the `_sumw2` array of the sparse bin with EFT weights.
 2. Filling a non-EFT sparse bin when `_sumw2 = None` resulted in the coffea `Hist._init_sumw2()` being called instead of `HistEFT._init_sum2()`.

The good news is that if you've never noticed these bugs, then they're irrelevant to you, because if you'd ever triggered on, it would result in a crash.  (That's what @btovar was seeing.)  In fact, I'm 99% sure that it's almost impossible to trigger (2) above the way our code was written, so fixing it didn't actually address @btovar's crash.  That required (1) which I actually figured out 2nd.

Anyway, I hope this can be expeditiously merged.  It only modifies `HistEFT.py` so it shouldn't interfere with the on-going event synchronization.